requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
chex==0.1.89
click==8.1.8
etils==1.12.2
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
flax==0.10.4
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
humanize==4.12.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.5.2
jaxlib==0.5.1
Jinja2==3.1.6
keras==3.9.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optax==0.2.4
optree==0.14.1
orbax-checkpoint==0.11.9
packaging==23.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_decision_forests==1.12.0
tensorflowjs==4.22.0
tensorstore==0.1.72
termcolor==2.5.0
tf_keras==2.19.0
toolz==1.0.0
treescope==0.1.9
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
wurlitzer==3.1.1
ydf==0.11.0
zipp==3.21.0
